% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateDistance.R
\name{calculateDistance}
\alias{calculateDistance}
\title{calculateDistance}
\usage{
calculateDistance(myData)
}
\arguments{
\item{myData}{data.frame with ship registers}
}
\value{
data.frame with ids of the observations (start and end), distance and difftime
}
\description{
The maximun distance between two consecutive observations (measured within three minutes)
}
